codeplayer-compiler-microservice/requirements.txt
epicbox
PyRSMQ
boto3
python-dotenv
python-socketio
simplejson